requirements.txt
django>=3.2
Pillow>=8.2.0
asgiref>=3.3.4
numpy>=1.20.2
opencv-python>=4.5.1.48
pip>=21.0.1
pytz>=2021.1
setuptools
sqlparse>=0.4.1
typing-extensions>=3.7.4.3
